Refactor logs
