+++ lib/logs/cubit/cubit.dart
@@ -1,0 +1,1 @@
+export 'log_cubit.dart';
