+++ lib/logs/cubit/log_state.dart
