+++ lib/logs/logs.dart
@@ -1,0 +1,4 @@
+export 'app_logger.dart';
+export 'cubit/cubit.dart';
+export 'log_file.dart';
+export 'log_page.dart';
