--- lib/presentation/logs/logs.dart
@@ -1,1 +1,0 @@
-export 'pages/log_page.dart';
