--- lib/presentation/logs/pages/pages.dart
@@ -1,1 +1,0 @@
-export 'log_page.dart';
